+++ test/example-types/dateRangeFacet.js
@@ -1,23 +1,28 @@
+let _ = require('lodash/fp')
 let { expect } = require('chai')
-let utils = require('./testUtils')
+let { getDateIfValid, rollingRangeToDates } = require('../../src/dateUtil')
 let dateRangeFacet = require('../../src/example-types/dateRangeFacet')
 
 let commonFilterParts = {
   type: 'dateRangeFacet',
   field: 'test',
   ranges: [
-    { from: 'now/d', key: 'open' },
-    { to: 'now-1d/d', key: 'expired' },
-    { from: 'now-30d/d', to: 'now/d', key: 'foo' },
+    { range: 'allFutureDates', key: 'open' },
+    { range: 'allPastDates', key: 'expired' },
   ],
 }
+
+let getDatePart = (range, part) => _.flow(
+  _.get(part),
+  getDateIfValid
+)(rollingRangeToDates(range, 'UTC'))
 
 describe('validContext', () => {
   it('should validate a context with all required fields', () => {
     expect(
       dateRangeFacet.validContext({
         field: 'value_field',
-        ranges: [{ key: 'asd', to: 'now/d', from: 'nod-1d/d' }],
+        ranges: [{ key: 'asd', range: 'allFutureDates' }],
       })
     ).to.be.true
   })
@@ -36,18 +41,10 @@ describe('validContext', () => {
       })
     ).to.be.false
   })
-  it('should invalidate a context with no range "from" or "to"', () => {
-    expect(
-      dateRangeFacet.validContext({
-        field: 'value_field',
-        ranges: [{ key: 'foo' }],
-      })
-    ).to.be.false
-  })
 })
 
 describe('dateRangeFacet/filter', () => {
-  it('should handle "from"', () => {
+  it('should handle "allFutureDates range"', () => {
     expect(
       dateRangeFacet.filter({
         ...commonFilterParts,
@@ -55,11 +52,17 @@ describe('dateRangeFacet/filter', () => {
       })
     ).to.deep.equal({
       bool: {
-        should: [{ range: { test: { from: 'now/d' } } }],
-      },
+        should: [{
+          range: {
+              test: {
+                from: getDatePart('allFutureDates', 'from')
+            }
+          }
+        }]
+      }
     })
   })
-  it('should handle "to"', () => {
+  it('should handle "allPastDates range"', () => {
     expect(
       dateRangeFacet.filter({
         ...commonFilterParts,
@@ -67,77 +70,14 @@ describe('dateRangeFacet/filter', () => {
       })
     ).to.deep.equal({
       bool: {
-        should: [{ range: { test: { to: 'now-1d/d' } } }],
-      },
-    })
-  })
-  it('should handle "from" and "to"', () => {
-    expect(
-      dateRangeFacet.filter({
-        ...commonFilterParts,
-        values: ['open', 'expired', 'foo'],
-      })
-    ).to.deep.equal({
-      bool: {
-        should: [
-          { range: { test: { from: 'now/d' } } },
-          { range: { test: { to: 'now-1d/d' } } },
-          { range: { test: { from: 'now-30d/d', to: 'now/d' } } },
-        ],
-      },
-    })
-  })
-})
-
-describe('dateRangeFacet/result', () => {
-  it('should work', () => {
-    utils.sequentialResultTest(
-      [
-        {
-          aggregations: {
-            range: {
-              buckets: [
-                {
-                  key: 'Expired',
-                  to: 1.5930432e12,
-                  to_as_string: '06-2020',
-                  doc_count: 6639865,
-                },
-                {
-                  key: 'Open',
-                  from: 1.5931296e12,
-                  from_as_string: '06-2020',
-                  doc_count: 75882,
-                },
-              ],
-            },
-          },
-        },
-      ],
-      {
-        field: 'test',
-        type: 'dateRangeFacet',
-        ranges: [
-          { from: 'now/d', key: 'Open' },
-          { to: 'now-1d/d', key: 'Expired' },
-        ],
-      },
-      [
-        { name: 'Open', count: 75882 },
-        { name: 'Expired', count: 6639865 },
-      ],
-      {
-        aggs: {
+        should: [{
           range: {
-            date_range: {
-              field: 'test',
-              format: 'MM-yyyy',
-              ranges: [{ from: 'now/d', key: 'Open' }],
-            },
-          },
-        },
-        size: 0,
+              test: {
+                to: getDatePart('allPastDates', 'to')
+            }
+          }
+        }]
       }
-    )
+    })
   })
 })
